V 0.1

+++ test/DaoTests/MysqlDaoDeleteTest.py
@@ -1,14 +1,8 @@
-import sys
-
-sys.path.insert(0, '/home/carana/PycharmProjects/test')
-
-from TransactionManager import TransactionManager
-from DatabasePersistence import DatabasePersistence
-from DatabaseDelegate import DatabaseDelegate
-from Model import Model
+from carlib.persistence.Constraints import Constraints
+from carlib.database.TransactionManager import TransactionManager
+from carlib.database.DatabasePersistence import DatabasePersistence
+from carlib.database.impl import MySQLBaseDelegate
 import logging
-
-from abc import ABCMeta
 
 logging.basicConfig(
     filename="test.log",
@@ -17,41 +11,21 @@ logging.basicConfig(
 )
 
 test = 'without_uid'
-#test = 'del_fk'
-#test = 'normal'
+# test = 'del_fk'
+# test = 'normal'
 
-class MySQLDAODelegate(DatabaseDelegate):
-    __metaclass__ = ABCMeta
-
-    def __init__(self):
-        DatabaseDelegate.__init__(self)
-
-    def is_duplicate_key_error(self, error_msg):
-        if error_msg.find("(23000): Duplicate entry") >= 0:
-            return True
-        return False
-
-    def is_foreign_key_error(self, error_msg):
-        # El primer chequeo es para odbc drivers , el otro es para pymssql
-        if (error_msg.find("conflicted with the FOREIGN KEY constraint") >= 0
-                or error_msg.find("a foreign key constraint fails") >= 0):
-            return True
-        return False
-
-    def get_uid(self, handler):
-        if hasattr(handler, 'lastrowid'):
-            return handler.lastrowid
-        return None
-
+calltest = True
 
 driver = 'mysql'
 trx = TransactionManager(driver, {'host': 'localhost', 'port': '3306',
                                   'user': 'root', 'password': 'melivane', 'database': 'py_dbtest'})
 
 
-class DAODelegateTest(MySQLDAODelegate):
+class DAODelegateTest(MySQLBaseDelegate):
     def __init__(self):
-        MySQLDAODelegate.__init__(self)
+        MySQLBaseDelegate.__init__(self)
+        if calltest:
+            self.delete_definition = {'is_call': True}
 
     def get_uid(self, cursor):
         # El stored procedure debe hacer solo insert a la tabla que se requeire el id o el insert debe ser
@@ -75,38 +49,54 @@ class DAODelegateTest(MySQLDAODelegate):
         elif test == 'del_fk':
             return "select * from tb_testfk where fk_test={}".format(key_values)
         else:
-            return "select * from tb_maintable_ckeys where main_code = '{}' and main_number = {}".format(key_values[0],key_values[1])
+            return "select * from tb_maintable_ckeys where main_code = '{}' and main_number = {}".format(key_values[0],
+                                                                                                         key_values[1])
 
-    def get_add_record_query(self, record_model, c_constraints=None, sub_operation=None):
-        return "insert into tb_maintable(anytext) values (DEFAULT,'{}') returning id_key".format(record_model.anytext)
-
-    def get_update_record_query(self, record_model, sub_operation=None):
-        return "update tb_maintable set anytext = '{}' where id_key = {}".format(record_model.anytext,record_model.id_key)
-
-    def get_delete_record_query(self, key_values):
+    def get_delete_record_query(self, key_values, c_constraints=None, sub_operation=None):
         if test == 'normal':
-            return "delete from tb_maintable where id_key = {}".format(key_values)
+            if calltest:
+                return 'dpdeletetest'
+            else:
+                return "delete from tb_maintable where id_key = {}".format(key_values)
         elif test == 'del_fk':
-            return "delete from tb_testfk where fk_test = {}".format(key_values)
+            if calltest:
+                return 'dpdeletetest_fk'
+            else:
+                return "delete from tb_testfk where fk_test = {}".format(key_values)
         else:
-            return "delete from tb_maintable_ckeys where main_code = '{}' and main_number = {}".format(key_values[0],key_values[1])
+            if calltest:
+                return 'dpdeletetest_ckeys'
+            else:
+                return "delete from tb_maintable_ckeys where main_code = '{}' and main_number = {}".format(
+                    key_values[0], key_values[1])
 
 
 daoDelegate = DAODelegateTest()
 
-# daoDelegate.set_add_definition()
-# daoDelegate.set_read_definition()
-
 dao = DatabasePersistence(trx, daoDelegate)
+
+constraints = None
 
 # usamos la transacion para informar que el control es extrerno.
 if test == 'normal':
-    ret = dao.delete_record(3)
+    key = 101
+    if calltest:
+        constraints = Constraints()
+        constraints.add_caller_parameter(Constraints.CallerOperation.DEL, key, 0)
+
+    ret = dao.delete_record(key, c_constraints=constraints)
 elif test == 'del_fk':
-    ret = dao.delete_record(1)
+    key = 1
+    if calltest:
+        constraints = Constraints()
+        constraints.add_caller_parameter(Constraints.CallerOperation.DEL, key, 0)
+    ret = dao.delete_record(key, c_constraints=constraints)
 else:
-    ret = dao.delete_record(('008',8))
+    keys = ('008', 8)
+    if calltest:
+        constraints = Constraints()
+        constraints.add_caller_parameter(Constraints.CallerOperation.DEL, keys[0], 0)
+        constraints.add_caller_parameter(Constraints.CallerOperation.DEL, keys[1], 1)
+    ret = dao.delete_record(keys, c_constraints=constraints)
+    # ret = dao.delete_record(None, c_constraints=constraints, verified_delete_check=False)
 print(ret)
-
-
-
